Fjerner splitting av tidslinjer. Må se om det blir bruk for det senere
