+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/eksempler/DifferanseBeregningTidslinje.kt
@@ -1,6 +1,6 @@
 package no.nav.familie.ba.sak.kjerne.tidslinje.eksempler
 
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.Tidslinje
+import no.nav.familie.ba.sak.kjerne.tidslinje.Tidslinje
 
 data class DifferanseBeregningTidslinje(
     val utbetalingTema: Tidslinje<Int>,
