+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/eksempler/KompetanseTidslinje.kt
@@ -1,18 +1,16 @@
 package no.nav.familie.ba.sak.kjerne.tidslinje.eksempler
 
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.Periode
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.TidslinjeMedEksterntInnhold
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.Tidsrom
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.rangeTo
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.tilTidspunktEllerUendeligLengeSiden
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.tilTidspunktEllerUendeligLengeTil
+import no.nav.familie.ba.sak.kjerne.tidslinje.Periode
+import no.nav.familie.ba.sak.kjerne.tidslinje.Tidslinje
+import no.nav.familie.ba.sak.kjerne.tidslinje.Tidsrom
+import no.nav.familie.ba.sak.kjerne.tidslinje.rangeTo
+import no.nav.familie.ba.sak.kjerne.tidslinje.tilTidspunktEllerUendeligLengeSiden
+import no.nav.familie.ba.sak.kjerne.tidslinje.tilTidspunktEllerUendeligLengeTil
 
 class KompetanseTidslinje(
     private val kompetanser: List<Kompetanse>,
-) : TidslinjeMedEksterntInnhold<Kompetanse>(
-    kompetanser
-) {
+) : Tidslinje<Kompetanse>() {
 
     override fun tidsrom(): Tidsrom {
         val fom = kompetanser.map { it.fom.tilTidspunktEllerUendeligLengeSiden { it.tom!! } }.minOrNull()!!
