+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/eksempler/KompetanseValideringTidslinje.kt
@@ -2,11 +2,11 @@ package no.nav.familie.ba.sak.kjerne.tidslinje.eksempler
 
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.KompetanseStatus
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.SnittTidslinje
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.Tidslinje
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.Tidspunkt
-import no.nav.familie.ba.sak.kjerne.eøs.temaperiode.hentUtsnitt
 import no.nav.familie.ba.sak.kjerne.tidslinje.PeriodeKombinator
+import no.nav.familie.ba.sak.kjerne.tidslinje.SnittTidslinje
+import no.nav.familie.ba.sak.kjerne.tidslinje.Tidslinje
+import no.nav.familie.ba.sak.kjerne.tidslinje.Tidspunkt
+import no.nav.familie.ba.sak.kjerne.tidslinje.hentUtsnitt
 
 enum class KompetanseValidering {
     OK_EØS_OG_KOMPETANSE,
